Fixup cc4c84e Upgraded snmp4j to another provider to avoid flaky tests
